+++ kite-core/src/main/kotlin/com/tang/kite/metadata/ColumnMeta.kt
@@ -22,7 +22,7 @@ class ColumnMeta(
     /**
      * Table name
      */
-    val tableName: String? = null,
+    val tableName: String,
 
     /**
      * Column name