+++ kite-core/src/main/kotlin/com/tang/kite/sql/ast/ddl/SqlStatementDdlHandler.kt
@@ -1,5 +1,6 @@
 package com.tang.kite.sql.ast.ddl
 
+import com.tang.kite.config.schema.SchemaConfig.getSql
 import com.tang.kite.metadata.ColumnMeta
 import com.tang.kite.sql.ast.AlterOperation
 import com.tang.kite.sql.ast.SqlNode
@@ -17,8 +18,8 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
         val sqlList = mutableListOf<String>()
         val sql = StringBuilder()
 
-        sql.append("create table ")
-        val tableName = createTableNode.table?.toString(false) ?: throw IllegalArgumentException("Table name cannot be null")
+        sql.append(getSql("create table "))
+        val tableName = createTableNode.table.toString()
         sql.append(tableName)
         sql.append(" (")
 
@@ -34,7 +35,7 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
         sql.append(")")
 
         if (createTableNode.comment != null && dialect.supportsCommentOnTable().not()) {
-            sql.append(" comment = '${createTableNode.comment}'")
+            sql.append(" ${getSql("comment")} = '${createTableNode.comment}'")
         }
         sqlList.add(sql.toString())
 
@@ -43,7 +44,7 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
         }
 
         if (createTableNode.comment != null && dialect.supportsCommentOnTable()) {
-            sqlList.add("comment on table ${createTableNode.table?.toString(false)} is '${createTableNode.comment}'")
+            sqlList.add("${getSql("comment on table")} ${createTableNode.table} ${getSql("is")} '${createTableNode.comment}'")
         }
 
         createTableNode.columns.forEach { column ->
@@ -57,41 +58,43 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
     override fun handleAlterTable(alterTableNode: SqlNode.AlterTable, dialect: SqlDialect): List<String> {
         val sqlList = mutableListOf<String>()
         val commentList = mutableListOf<String>()
-        val tableName = alterTableNode.table?.toString(false) ?: throw IllegalArgumentException("Table name cannot be null")
+        val tableName = alterTableNode.table.toString()
 
         alterTableNode.operations.forEach { operation ->
             val sql = StringBuilder()
-            sql.append("alter table ").append(tableName).append(" ")
+            sql.append(getSql("alter table ")).append(tableName).append(" ")
 
             when (operation) {
                 is AlterOperation.AddColumn -> {
-                    sql.append(dialect.getAddColumnKeyword()).append(" ")
+                    sql.append(getSql(dialect.getAddColumnKeyword())).append(" ")
                     appendColumnMeta(sql, operation.column, dialect)
                     commentList.add(getColumnComment(tableName, operation.column, dialect))
                 }
                 is AlterOperation.DropColumn -> {
-                    sql.append(dialect.getDropColumnKeyword())
+                    sql.append(getSql(dialect.getDropColumnKeyword()))
                     sql.append(" ${operation.columnName}")
                     if (operation.cascade && dialect.supportsCascade()) {
-                        sql.append(" cascade")
+                        sql.append(getSql(" cascade"))
                     }
                 }
                 is AlterOperation.ModifyColumn -> {
-                    sql.append(dialect.getAlterColumnKeyword()).append(" ")
+                    sql.append(getSql(dialect.getAlterColumnKeyword())).append(" ")
                     appendColumnMeta(sql, operation.column, dialect)
                     commentList.add(getColumnComment(tableName, operation.column, dialect))
                 }
                 is AlterOperation.RenameColumn -> {
-                    sql.append("rename column ${operation.oldName} to ${operation.newName}")
+                    sql.append(getSql("rename column "))
+                    sql.append("${operation.oldName} ${getSql("to")} ${operation.newName}")
                 }
                 is AlterOperation.AddConstraint -> {
-                    sql.append("add ")
+                    sql.append(getSql("add "))
                     appendTableConstraint(sql, operation.constraint)
                 }
                 is AlterOperation.DropConstraint -> {
-                    sql.append("drop constraint ${operation.constraintName}")
+                    sql.append(getSql("drop constraint "))
+                    sql.append(operation.constraintName)
                     if (operation.cascade && dialect.supportsCascade()) {
-                        sql.append(" cascade")
+                        sql.append(getSql(" cascade"))
                     }
                 }
             }
@@ -103,29 +106,29 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
 
     override fun handleDropTable(dropTableNode: SqlNode.DropTable, dialect: SqlDialect): List<String> {
         val sql = StringBuilder()
-        val tableName = dropTableNode.table?.toString(false) ?: throw IllegalArgumentException("Table name cannot be null")
-        sql.append("drop table ")
+        val tableName = dropTableNode.table.toString()
+        sql.append(getSql("drop table "))
         sql.append(tableName)
         if (dropTableNode.cascade && dialect.supportsCascade()) {
-            sql.append(" cascade")
+            sql.append(getSql(" cascade"))
         }
         return listOf(sql.toString())
     }
 
     override fun handleCreateIndex(createIndexNode: SqlNode.CreateIndex, dialect: SqlDialect): List<String> {
         val sql = StringBuilder()
-        sql.append("create ")
+        sql.append(getSql("create "))
         if (createIndexNode.unique) {
-            sql.append("unique ")
+            sql.append(getSql("unique "))
         }
-        sql.append("index ")
+        sql.append(getSql("index "))
         sql.append(createIndexNode.indexName)
-        sql.append(" on ")
-        sql.append(createIndexNode.table.toString(false))
+        sql.append(getSql(" on "))
+        sql.append(createIndexNode.table.toString())
         sql.append(" (")
         val columns = createIndexNode.columns.mapIndexed { index, string ->
             val order = createIndexNode.sorts.getOrNull(index)
-            if (order != null) "$string $order" else string
+            if (order != null) "$string ${getSql(order.name)}" else string
         }
         sql.append(columns.joinToString(", "))
         sql.append(")")
@@ -134,10 +137,10 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
 
     override fun handleDropIndex(dropIndexNode: SqlNode.DropIndex, dialect: SqlDialect): List<String> {
         val sql = StringBuilder()
-        sql.append("drop index ")
+        sql.append(getSql("drop index "))
         sql.append(dropIndexNode.indexName)
         if (dropIndexNode.table != null && dialect.requiresTableForDropIndex()) {
-            sql.append(" on ")
+            sql.append(getSql(" on "))
             sql.append(dropIndexNode.table.toString())
         }
         return listOf(sql.toString())
@@ -145,10 +148,10 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
 
     override fun handleTruncateTable(truncateTableNode: SqlNode.TruncateTable, dialect: SqlDialect): List<String> {
         val sql = StringBuilder()
-        sql.append("truncate table ")
-        sql.append(truncateTableNode.table?.toString(false))
+        sql.append(getSql("truncate table "))
+        sql.append(truncateTableNode.table.toString())
         if (truncateTableNode.cascade && dialect.supportsCascade()) {
-            sql.append(" cascade")
+            sql.append(getSql(" cascade"))
         }
         return listOf(sql.toString())
     }
@@ -158,23 +161,25 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
         sql.append(" ")
         sql.append(getType(column))
         if (column.nullable.not()) {
-            sql.append(" not null")
+            sql.append(getSql(" not null"))
         }
         if (column.defaultValue.isNullOrBlank().not()) {
-            sql.append(" default ${column.defaultValue}")
+            sql.append(getSql(" default "))
+            sql.append(column.defaultValue)
         }
         if (column.autoIncrement) {
             sql.append(" ")
             sql.append(dialect.getAutoIncrementKeyword())
         }
         if (column.primaryKey) {
-            sql.append(" primary key")
+            sql.append(getSql(" primary key"))
         }
         if (column.unique) {
-            sql.append(" unique")
+            sql.append(getSql(" unique"))
         }
         if (column.comment != null && dialect.supportsCommentOnColumn().not()) {
-            sql.append(" comment '${column.comment}'")
+            sql.append(getSql(" comment "))
+            sql.append("'${column.comment}'")
         }
     }
 
@@ -182,52 +187,54 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
         when (constraint) {
             is TableConstraint.PrimaryKey -> {
                 if (constraint.name != null) {
-                    sql.append("constraint ${constraint.name} ")
+                    sql.append("${getSql("constraint")} ${constraint.name} ")
                 }
-                sql.append("primary key (${constraint.columns.joinToString(", ")})")
+                sql.append("${getSql("primary key")} (${constraint.columns.joinToString(", ")})")
             }
             is TableConstraint.Unique -> {
                 if (constraint.name != null) {
-                    sql.append("constraint ${constraint.name} ")
+                    sql.append("${getSql("constraint")} ${constraint.name} ")
                 }
-                sql.append("unique (${constraint.columns.joinToString(", ")})")
+                sql.append("${getSql("unique")} (${constraint.columns.joinToString(", ")})")
             }
             is TableConstraint.ForeignKey -> {
                 if (constraint.name != null) {
-                    sql.append("constraint ${constraint.name} ")
+                    sql.append("${getSql("constraint")} ${constraint.name} ")
                 }
-                sql.append("foreign key (${constraint.columns.joinToString(", ")}) ")
-                sql.append("references ${constraint.referencedTable} (${constraint.referencedColumns.joinToString(", ")})")
+                sql.append("${getSql("foreign key")} (${constraint.columns.joinToString(", ")}) ")
+                sql.append("${getSql("references")} ${constraint.referencedTable} (${constraint.referencedColumns.joinToString(", ")})")
 
                 // Add foreign key actions if specified
                 if (constraint.onDelete != null) {
-                    sql.append(" on delete ${constraint.onDelete.name.lowercase().replace("_", " ")}")
+                    sql.append(getSql(" on delete "))
+                    sql.append(constraint.onDelete.name.lowercase().replace("_", " "))
                 }
                 if (constraint.onUpdate != null) {
-                    sql.append(" on update ${constraint.onUpdate.name.lowercase().replace("_", " ")}")
+                    sql.append(getSql(" on update "))
+                    sql.append(constraint.onUpdate.name.lowercase().replace("_", " "))
                 }
             }
             is TableConstraint.Check -> {
                 if (constraint.name != null) {
-                    sql.append("constraint ${constraint.name} ")
+                    sql.append("${getSql("constraint")} ${constraint.name} ")
                 }
-                sql.append("check (${constraint.expression})")
+                sql.append("${getSql("check")} (${constraint.expression})")
             }
         }
     }
 
     private fun getType(column: ColumnMeta): String {
         val sql = StringBuilder()
-        sql.append(column.typeName)
+        sql.append(getSql(column.typeName))
         when {
-            column.typeName.startsWith("decimal") || column.typeName.startsWith("numeric") -> {
+            column.typeName.startsWith("decimal", ignoreCase = true) || column.typeName.startsWith("numeric", ignoreCase = true) -> {
                 if (column.columnSize > 0 && column.decimalDigits >= 0) {
                     sql.append("(${column.columnSize}, ${column.decimalDigits})")
                 } else {
                     sql.append("(10, 2)")
                 }
             }
-            column.typeName.startsWith("varchar") || column.typeName.startsWith("char") -> {
+            column.typeName.startsWith("varchar", ignoreCase = true) || column.typeName.startsWith("char", ignoreCase = true) -> {
                 if (column.columnSize > 0) {
                     sql.append("(${column.columnSize})")
                 } else {
@@ -241,12 +248,12 @@ object SqlStatementDdlHandler : DdlHandler<List<String>> {
     fun getColumnComment(tableName: String, column: ColumnMeta, dialect: SqlDialect): String {
         val columnName = column.columnName
         val sql = if (dialect.supportsCommentOnColumn()) {
-            "comment on column $tableName.$columnName is '${column.comment ?: ""}'"
+            "${getSql("comment on column")} $tableName.$columnName ${getSql("is")} '${column.comment ?: ""}'"
         } else {
             if (column.comment == null) {
-                "drop comment on column $tableName.$columnName"
+                "${getSql("drop comment on column")} $tableName.$columnName"
             } else {
-                "alter table $tableName modify column $columnName ${getType(column)} comment '${column.comment}'"
+                "${getSql("alter table")} $tableName ${getSql("modify column")} $columnName ${getType(column)} ${getSql("comment")} '${column.comment}'"
             }
         }
         return sql